task22/neuro_control/frontend/requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
